added data and watering endpoints for pi
