+++ server.js
@@ -1,14 +1,42 @@
-const { createServer } = require('node:http');
+const express = require('express');
+const cookieParser = require('cookie-parser');
 
-const port = process.env.PORT || 3000; // Use Azure's PORT or default to 3000
+const app = express();
+const port = process.env.PORT; 
 
-const server = createServer((req, res) => {
-  res.statusCode = 200;
-  res.setHeader('Content-Type', 'text/plain');
-  res.end('Hello World');
+app.use(cookieParser());
+app.use(express.static('public'));
+app.use(express.json());
+
+const mainController = require('./controllers/mainController');
+
+app.set('view engine', 'ejs');
+app.set('views', './views');
+
+// Get Routes
+
+app.get('/', mainController.getIndex);
+app.get('/product', mainController.getProduct);
+app.get('/home', mainController.getHome);
+app.get('/checkout', mainController.getCheckout);
+app.get('/login', mainController.getLogin);
+app.get('/product-admin', mainController.getProductAdmin);
+
+// Post Routes
+
+app.post('/addToCart', mainController.addToCart);
+app.post('/removeFromCart', mainController.removeFromCart);
+app.post('/emptyCart', mainController.emptyCart);
+app.post('/showToProducts', mainController.showToProducts)
+app.post('/hideFromProducts', mainController.hideFromProducts)
+app.post('/addToProducts', mainController.addToProducts)
+app.post('/addUser', mainController.addUser);
+app.post('/verifyUser', mainController.verifyUser);
+app.post('/addQuantity', mainController.addQuantity);
+app.post('/productRedirect', function (req, res) {
+    res.redirect(303, '/product');
 });
 
-
-server.listen(port, '0.0.0.0', () => {
-  console.log(`Server running on port ${port}`);
+app.listen(port, () => {
+    console.log(`Server is running on http://localhost:${port}`);
 });
